Added infrastructure and scripts for automatically scaffolding new subjects and transferring fmri data into their neuropipe folder

+++ scripts/move-behavioral-data-all-subjs.sh
@@ -8,5 +8,8 @@ source globals.sh
 
 for subj in $ALL_SUBJECTS; do
 	mkdir -p ./subjects/$subj/data/session_info/
-	cp ${DATA_DIR}/session_log/$subj/log.txt ./subjects/$subj/data/session_info/
+	# only copy it over if it doesn't already exist
+	if [[ ! -f "./subjects/$subj/data/session_info/log.txt" ]]; then
+		cp ${DATA_DIR}/session_log/$subj/log.txt ./subjects/$subj/data/session_info/
+	fi
 done
